Remove `FeedPhoto` reference from `FeedPhotosMapper`
