+++ photofeed/src/commonMain/kotlin/com/asturiancoder/photofeed/feed/api/RemoteFeedRepository.kt
@@ -1,7 +1,9 @@
 package com.asturiancoder.photofeed.feed.api
 
+import com.asturiancoder.photofeed.feed.api.model.RemoteFeedPhoto
 import com.asturiancoder.photofeed.feed.feature.FeedLoader
 import com.asturiancoder.photofeed.feed.feature.FeedPhoto
+import com.asturiancoder.photofeed.util.Uuid
 
 class RemoteFeedRepository(
     private val url: String,
@@ -21,9 +23,26 @@ class RemoteFeedRepository(
         }
 
         return try {
-            FeedPhotosMapper.map(response)
+            FeedPhotosMapper.map(response).toModel()
         } catch (exception: Exception) {
             throw Error.InvalidData
         }
     }
 }
+
+private fun List<RemoteFeedPhoto>.toModel(): List<FeedPhoto> =
+    mapNotNull { remotePhoto ->
+        Uuid.from(remotePhoto.id)?.let { uuid ->
+            FeedPhoto(
+                id = uuid,
+                description = remotePhoto.description,
+                location = remotePhoto.location,
+                url = remotePhoto.url,
+                likes = remotePhoto.likes,
+                author = FeedPhoto.Author(
+                    name = remotePhoto.author.name,
+                    imageUrl = remotePhoto.author.imageUrl,
+                ),
+            )
+        }
+    }
